+++ lib/libalpm/remove.c
@@ -629,7 +629,8 @@ static int remove_package_files(alpm_handle_t *handle,
 		}
 	}
 
-	_alpm_log(handle, ALPM_LOG_DEBUG, "removing %zu files\n", filelist->count);
+	_alpm_log(handle, ALPM_LOG_DEBUG, "removing %llu files\n",
+			(unsigned long long)filelist->count);
 
 	if(!newpkg) {
 		/* init progress bar, but only on true remove transactions */
